geodetic_calculator/requirements.txt
Django==5.0.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
pandas==2.2.3
openpyxl==3.1.5
pyproj==3.7.1
